Add files via upload
The day light system works fine, but the layer portion might not, it might be affecting the building portion but Im not completely sure. Depending on the color of the current cost, it will change the texture of the ground. And sfx is there but I also can not text weather it works currently or not.